requirements.txt
dwavebinarycsp
dwave-system
matplotlib
dwave-inspector